yoloxdetector/yolox_streamlit_app/requirements.txt
streamlit==1.45.1
moviepy==2.2.0
imageio==2.37.0
imageio-ffmpeg==0.6.0
pytube==15.0.0
yt-dlp==2025.4.30
proglog==0.1.12
pillow==11.2.1
opencv-python==4.11.0.86
numpy==2.2.6
tabulate==0.9.0
psutil==7.0.0